feat: workspace complete
